requirements.txt
tiatoolbox==1.6.0
timm==1.0.11
ttach==0.0.3
segmentation-models-pytorch==0.3.4
numcodecs==0.11.0

